Fix height bug in loginpage and refactor to vanilla css
