+++ apps/store/src/pages/[locale]/member/login.tsx
@@ -1,6 +1,6 @@
 import type { GetServerSidePropsContext } from 'next'
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
-import { LoginPage } from '@/features/memberArea/pages/LoginPage'
+import { LoginPage } from '@/features/memberArea/pages/LoginPage/LoginPage'
 import { getAccessToken } from '@/services/authApi/persist'
 import { isRoutingLocale } from '@/utils/l10n/localeUtils'
 import { ORIGIN_URL } from '@/utils/PageLink'
